+++ components/tvplanit/source/vpflxds.pas
@@ -161,7 +161,7 @@ type
     function GetFieldName(Mappings: TCollection;                         
       VPField: string): string;                                          
 
-    function GetNextID(TableName: string): Integer; override;
+    function GetNextID(TableName: string): Int64; override;
 
     { These are published via the TVpDataSources class, which allows them to }
     { be clustered in the Object Inspector Don't publish them individually   }
@@ -2031,7 +2031,7 @@ begin
 end;
 {=====}
 
-function TVpFlexDataStore.GetNextID(TableName: string): Integer;
+function TVpFlexDataStore.GetNextID(TableName: string): Int64;
 begin
   { The FlexDataStore has no idea what type of database you are connected to }
   { beyond TDataset compatibility, so it cannot presume to generate record   }
